refactor: move console/index.ts to console.ts
