+++ packages/log/ts/logStore.ts
@@ -1,6 +1,6 @@
 
 import { mapKey, reduceByKey, reduceKey } from 'type-plus'
-import { createConsoleLogReporter } from './console/index.js'
+import { createConsoleLogReporter } from './console.js'
 import { logLevels, toLogLevel, toLogLevelName } from './logLevels.js'
 import { createColorLogReporter } from './platform/index.js'
 import type { Logger, LogLevel, LogReporter, StandardLogOptions } from './types.js'
